Rename a few files and clean up wording.
