+++ ql/src/experimental/IntegerOverflow/IntegerOverflowGood.go
@@ -1,0 +1,8 @@
+package main
+
+func getSubSliceGood(buf []byte, start int, offset int) []byte {
+	if start+offset < start {
+		return nil
+	}
+	return buf[start : start+offset]
+}
