--- ql/src/experimental/integer-overflow-detection/integer-overflow-solution-example.go
@@ -1,6 +1,0 @@
-func getSubSlice(buf []byte, start int, offset int) []byte {
-	if(start + offset < start) {
-		return nil
-	}
-	return buf[start:start + offset]
-}
